Introduce the BOX method for boxing objects up.
This change gets rid of the notion that strings and characters can be
boxes. This is because that will hamper editing later on.

Also introduce WITH-RECORDING-OFF to disable region recording within a
particular context. *FORCE-RECORDING* will force the recording of
regions anyway for debugging, but this is disabled by default.